.history/backend/tests/requirements-test_20250330202452.txt
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
requests==2.31.0
coverage==7.3.2
redis==5.0.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-RateLimiter==0.1.0
graphene==3.3
graphene-sqlalchemy==3.0.1
marshmallow==3.20.1
structlog==23.2.0
unidecode==1.3.7 